[Lighting] Fixed a bug where the regular radial distance fog wasn't visible

+++ runtime/Core/Timer.cpp
@@ -53,7 +53,7 @@ namespace Spartan
 
     void Timer::Initialize()
     {
-        fps_limit = Display::GetRefreshRate();
+        fps_limit = static_cast<uint32_t>(Display::GetRefreshRate());
     }
 
     void Timer::PostTick()
